+++ src/main/java/com/codetaylor/mc/pyrotech/modules/core/item/spi/ItemHammerBase.java
@@ -1,5 +1,6 @@
 package com.codetaylor.mc.pyrotech.modules.core.item.spi;
 
+import com.codetaylor.mc.pyrotech.modules.core.ModuleCoreConfig;
 import net.minecraft.block.Block;
 import net.minecraft.client.util.ITooltipFlag;
 import net.minecraft.item.ItemStack;
@@ -22,7 +23,7 @@ public abstract class ItemHammerBase
   @Override
   public void addInformation(ItemStack stack, @Nullable World world, List<String> tooltip, ITooltipFlag flag) {
 
-    if (this.getDamage(stack) == 0) {
+    if (ModuleCoreConfig.CLIENT.SHOW_DURABILITY_TOOLTIPS && this.getDamage(stack) == 0) {
       tooltip.add(I18n.translateToLocalFormatted("gui.pyrotech.tooltip.durability.full", this.getMaxDamage(stack)));
     }
   }