+++ src/main/java/com/codetaylor/mc/pyrotech/modules/ignition/item/ItemFlintAndTinder.java
@@ -1,5 +1,6 @@
 package com.codetaylor.mc.pyrotech.modules.ignition.item;
 
+import com.codetaylor.mc.pyrotech.modules.core.ModuleCoreConfig;
 import com.codetaylor.mc.pyrotech.modules.ignition.ModuleIgnitionConfig;
 import net.minecraft.client.util.ITooltipFlag;
 import net.minecraft.entity.EntityLivingBase;
@@ -41,7 +42,7 @@ public class ItemFlintAndTinder
   @Override
   public void addInformation(ItemStack stack, @Nullable World world, List<String> tooltip, ITooltipFlag flag) {
 
-    if (this.getDamage(stack) == 0) {
+    if (ModuleCoreConfig.CLIENT.SHOW_DURABILITY_TOOLTIPS && this.getDamage(stack) == 0) {
       tooltip.add(I18n.translateToLocalFormatted("gui.pyrotech.tooltip.durability.full", this.getMaxDamage(stack)));
     }
   }