+++ src/main/java/com/codetaylor/mc/pyrotech/modules/tool/item/spi/ItemPickaxeBase.java
@@ -1,5 +1,6 @@
 package com.codetaylor.mc.pyrotech.modules.tool.item.spi;
 
+import com.codetaylor.mc.pyrotech.modules.core.ModuleCoreConfig;
 import com.codetaylor.mc.pyrotech.modules.tool.ModuleToolConfig;
 import net.minecraft.client.util.ITooltipFlag;
 import net.minecraft.item.ItemPickaxe;
@@ -34,7 +35,7 @@ public class ItemPickaxeBase
   @Override
   public void addInformation(ItemStack stack, @Nullable World world, List<String> tooltip, ITooltipFlag flag) {
 
-    if (this.getDamage(stack) == 0) {
+    if (ModuleCoreConfig.CLIENT.SHOW_DURABILITY_TOOLTIPS && this.getDamage(stack) == 0) {
       tooltip.add(I18n.translateToLocalFormatted("gui.pyrotech.tooltip.durability.full", this.getMaxDamage(stack)));
     }
   }